+++ CHANGES.rst
@@ -112,6 +112,9 @@ Other Changes and Additions
 Bug Fixes
 ---------
 
+- Avoid a non-finite error in model fitting by not passing spectrum uncertainties as
+  weights if the uncertainty values are all 0. [#1880]
+
 Cubeviz
 ^^^^^^^